apps/recruitment/jobhax/job-application-platform/requirements.txt
# Core dependencies
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# AI/ML libraries
openai>=1.3.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Web scraping and automation
playwright>=1.40.0
webdriver-manager>=4.0.0
fake-useragent>=1.4.0

# Image processing (for CV analysis)
Pillow>=10.0.0
opencv-python>=4.8.0

# PDF processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
